Add files via upload
The source codes for recommendation generation are in py files named conten_based, movie_emotion_based and movie_genres_based. The app.py is the integrating flask file and remaining files are the raw web page codes for recommendation systems and output webpages. For better understanding, go through the flask main page named app.py to understand the mapping of all the files. The movie_dataset is collected from kaggle due to un-availability of live data.